refactor: type 정리 및 export
